change implementation of SOS-Theory
The existing implementation of SOS-theory treats theory-axiom clauses
as SOS-clauses. It assumes that theory axioms are already added to the
search space as clauses, and in particular have an inference which
satisfies isTheoryAxiom().
On the one hand, this property does currently not hold, as there are
theory axioms which are added as formulas and not as clauses (e.g.
the exhaustiveness axiom for term algebras). Such axioms are then
clausified, and the resulting clauses are derived by the
cnf-transformation and therefore do not satisfy isTheoryAxiom().
On the other hand, we want to allow the user to specify theory axioms
externally. Such axioms will be added as formulas (in particular if
SMTLIB-input is used, which does not have a notion of clausal input),
and the resulting clauses will again not satisfy isTheoryAxiom().

As a solution, we replace for SOS-theory the call to isTheoryAxiom() with
isPureTheoryDescendant(). This way, clauses, which result from
clausifying theory axioms will also be captured, additionally to clausal
theory axioms.